Decrease API requests
A couple of changes.
First of all, we've removed the part that for each
track entry requests the display name of the user that added the entry.
Instead we're now adding the ids into a list, that we will then go over
and get the display names of each id (probably put them in a dict).
Secondly, we're now keeping track of the requests we make and if we make
too many, we're waiting 30 seconds. I might have to tweak how often it
waits and the sort but for now this is what we have.