Fixes scope issue
Ok. Turns out I deleted one of the necessary scopes and due to the fact
I had already allowed it previously it didn't register as an error, but
cloning the project elsewhere very much did register the error. But this
time, it really is the end, for reals. I hope.